Move checker to dynamo package
