+++ core/src/main/java/com/scalar/db/storage/dynamo/DynamoOperationChecker.java
@@ -1,16 +1,20 @@
-package com.scalar.db.storage.common.checker;
+package com.scalar.db.storage.dynamo;
 
 import com.scalar.db.api.Put;
 import com.scalar.db.api.TableMetadata;
 import com.scalar.db.common.TableMetadataManager;
+import com.scalar.db.storage.common.checker.ColumnChecker;
+import com.scalar.db.storage.common.checker.OperationChecker;
+import javax.annotation.concurrent.ThreadSafe;
 
+@ThreadSafe
 public class DynamoOperationChecker extends OperationChecker {
   public DynamoOperationChecker(TableMetadataManager metadataManager) {
     super(metadataManager);
   }
 
   @Override
-  void checkColumnsInPut(Put put, TableMetadata metadata) {
+  protected void checkColumnsInPut(Put put, TableMetadata metadata) {
     super.checkColumnsInPut(put, metadata);
 
     put.getColumns().values().stream()
